remove base_inferencer
