+++ mmaction/apis/inferencers/__init__.py
@@ -1,8 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .actionrecog_inferencer import ActionRecogInferencer
-from .base_mmaction_inferencer import BaseMMAction2Inferencer
 from .mmaction_inferencer import MMAction2Inferencer
 
-__all__ = [
-    'ActionRecogInferencer', 'BaseMMAction2Inferencer', 'MMAction2Inferencer'
-]
+__all__ = ['ActionRecogInferencer', 'MMAction2Inferencer']
